backend/pyproject.toml
[project]
name = "prod"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "strawberry-graphql[fastapi] (>=0.261.1,<0.262.0)",
    "sqlmodel (>=0.0.23,<0.0.24)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "redis (>=5.2.1,<6.0.0)",
    "sentence-transformers (>=3.4.1,<4.0.0)",
    "aiobotocore (>=2.21.0,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"
flake8 = "^7.1.2"
black = "^25.1.0"
pre-commit = "^4.1.0"
icecream = "^2.1.4"
mypy = "^1.15.0"



[tool.poetry.group.tests.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.25.3"

[tool.flake8]
ignore = ["D203", "E501"]
exclude = [".git", "__pycache__", "__init__.py", "docs/"]
max-complexity = 18
max-line-length = 80

[tool.isort]
profile = "black"
default_section = "THIRDPARTY"
balanced_wrapping = true
known_first_party = "src"
line_length = 80
lines_after_imports = 2
lines_between_sections = 1
multi_line_output = 3
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
skip = ["env", ".env", ".env.example", "migrations/"]

[tool.black]
line-length = 80
skip-string-normalization = false

[tool.mypy]
disable_error_code = ["no-redef", "import-not-found", "import-untyped", "attr-defined"]

